Refactor IPC into modular handlers
